% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonlat.R
\docType{data}
\name{lonlat}
\alias{lonlat}
\title{Coordinates for Entire Brazil}
\format{
\subsection{\code{lonlat}}{

A matrix with 2841 rows and 2 columns:
\describe{
\item{lon}{Longitude in decimal degrees}
\item{lat}{Latitude in decimal degrees}
...
}
}
}
\source{
Data generated using information from the Instituto Brasileiro de Geografia e Estatística (IBGE).
For more information, visit: \url{https://www.ibge.gov.br/}
}
\usage{
lonlat
}
\description{
This dataset provides coordinates for Brazil at a spatial resolution of 0.5 x 0.5 degrees.
}
\keyword{datasets}
