% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OxfordRain.R
\docType{data}
\name{OxfordRain}
\alias{OxfordRain}
\title{Daily Rainfall Amounts}
\format{
A data frame with 70523 rows and 1 column:
\describe{
\item{Rain}{Precipitation in millimeters}
}
}
\source{
<doi.org/10.6084/m9.figshare.11956239.v1>
}
\usage{
OxfordRain
}
\description{
Daily rainfall data from Oxford-UK (Radcliffe Observatory).
Covers the period from January 1827 to January 2020.
Belongs to Burt, Stephen (2020). Daily meteorological data from
the Radcliffe Observatory (now Radcliffe Meteorological Station),
from January 1815: updated to January 2020. figshare. Dataset.
}
\keyword{datasets}
