% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainTS4.R
\docType{data}
\name{rainTS4}
\alias{rainTS4}
\title{Rainfall Aggregated at the 4-quasi week time scale}
\format{
\subsection{\code{rainTS4}}{

A matrix with 4 columns and 576 rows:
\describe{
\item{Year}{Year}
\item{Month}{Month}
\item{quasiWeek}{Four quasi-weekly periods in each month}
\item{rain.at.TS4}{Rainfall amounts aggregated at TS equal to four}
...
}
}
}
\source{
\url{https://psl.noaa.gov/}
}
\usage{
rainTS4
}
\description{
Rainfall amounts aggregated using \code{TSaggreg()}. Campinas state of Sao Paulo, Brazil.
}
\keyword{datasets}
