% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CampinasRain.R
\docType{data}
\name{CampinasRain}
\alias{CampinasRain}
\title{Daily Rainfall Amounts}
\format{
\subsection{\code{CampinasRain}}{

A dataframe with 2 columns and 16071 rows:
\describe{
\item{Date}{Date}
\item{Rain}{Daily rainfall amounts recorded in millimetre in Campinas state of Sao Paulo, Brazil}
...
}
}
}
\source{
\url{https://psl.noaa.gov/}
}
\usage{
CampinasRain
}
\description{
Daily rainfall amounts recorded in millimetre in Campinas state of Sao Paulo, Brazil.
}
\keyword{datasets}
